finished typisifying
