+++ app/(auth)/_layout.tsx
@@ -1,15 +1,13 @@
+import { StyleSheet, Text, View } from 'react-native'
 import {SplashScreen, Stack, Tabs, Redirect} from 'expo-router';
 import {useFonts} from 'expo-font'
-
-
-
 
 const AuthLayout = () => {
 
   const [fontsLoaded, error] = useFonts({
-    "Odin" : require("../../assets/fonts/Odin-Bold.otf"),
-    "Inter" : require("../../assets/fonts/Inter Regular.ttf")
-})
+        "Odin" : require("../../assets/fonts/Odin-Bold.otf"),
+        "Inter" : require("../../assets/fonts/Inter Regular.ttf")
+    })
 
 return (
   <Stack>
