+++ app/(tabs)/favorites.tsx
