+++ components/mall.tsx
@@ -3,8 +3,8 @@ import NextArrow from '../assets/images/tabs/mall/next.png'
 import StoreSideScroll from './storeSideScroll';
 
 export interface MallProps{
-    data: { id: number, storeCat: number, promo: string, closed: string, icon: string, name: string, openT: String, closeT: String }[],
-    catData: { id: number, uid: number, name: string }[]
+    data: { id: number, storeCat: string, promo?: string, closed: boolean, icon: string, name: string, openT: String, closeT: String }[],
+    catData: { id: number, uid: string, name: string }[]
 }
 
 const Mall: React.FC<MallProps> = ({data, catData}) => {
