+++ lib/onbData.ts
@@ -3,21 +3,21 @@ export const onbData = [
         id: 1,
         header: 'Welcome to SKG Mall!',
         des: 'Order from your favorite restaurants, Explore \n different flavors and Satisfy your cravings \n with ease.',
-        image: '../assets/images/onb/boy eating illustration.png',
+        image: require('../assets/images/onb/boy eating illustration.png'),
         size: 'h-[250px] w-[164.2px]'
     },
     {
         id: 2,
         header: 'Your Favorite Eateries, All in One Place.',
         des: 'Delve into a diverse selection of culinary delights \n from top-rated restaurants.',
-        image: '../assets/images/onb/hamburger illustration.png',
+        image: require('../assets/images/onb/hamburger illustration.png'),
         size: 'h-[250px] w-[297.4px]'
     },
     {
         id: 3,
         header: 'Effortless Ordering',
         des: 'Indulge in the ease of our user-friendly ordering system, guaranteeing prompt deliveries and \n a hassle-free dining experience every time.',
-        image: '../assets/images/onb/order food illustration.png',
+        image: require('../assets/images/onb/order food illustration.png'),
         size: 'h-[250px] w-[260.89px]'
     },
